backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
requests>=2.31.0
beautifulsoup4>=4.12.2
cohere>=4.0.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0
openai>=1.3.5
google-generativeai>=0.4.0
pydantic-settings>=2.0.0
pytest>=7.4.0